Fix and reorganised schema tests and fixtures #21
